Revamp all data building and checking
Thorough use of assertr to check that values are present and unique when they ought to be. Also, dt_compiled does all the joining and spreading work necessary to create the real training data for the model. Filtered subsets of this main table can then be more easily used in modeling code - with 80% of the data carpentry finished ahead of time.